requirements.txt
django==2.0.3
pymysql==0.8.0
Pillow==5.0.0
argparse==1.4.0
django-debug-toolbar==1.9.1
django-filter==1.1
djangorestframework==3.7.7
easy-thumbnails==2.5
pygeocoder==1.2.5
requests==2.18.4
sqlparse==0.2.4
